Refactor(OutputHandling): Create array size and allocator classes
Created a handful of size classes that get 1d, 2d, or 3d array sizes based on different-system-level inputs. Added a BatchArrays class which will cache and allocate the output arrays for a full batch to reduce allocation overhead. Scrubbed old tests for a previous version. Old code left in comments while development of the allocator continues.

ref: #17, #45, #44, #39, #35, #31, #23, #22, #15, #11, #1